requirements.txt
asgiref==3.7.2
crispy-bootstrap4==2022.1
Django==4.2.3
django-crispy-forms==2.0
django-extensions==3.2.3
django-phonenumber-field==7.1.0
django-taggit==4.0.0
djangorestframework==3.14.0
phonenumberslite==8.13.17
Pillow==10.0.0
psycopg2==2.9.6
python-dotenv==1.0.0
pytz==2023.3
sqlparse==0.4.4
sweetify==2.3.1
tzdata==2023.3
